+++ src/SFML/Window/Joystick.cpp
@@ -50,7 +50,7 @@ unsigned int Joystick::getButtonCount(unsigned int joystick)
 ////////////////////////////////////////////////////////////
 bool Joystick::hasAxis(unsigned int joystick, Axis axis)
 {
-    return priv::JoystickManager::getInstance().getCapabilities(joystick).axes[static_cast<int>(axis)];
+    return priv::JoystickManager::getInstance().getCapabilities(joystick).axes[axis];
 }
 
 
@@ -65,7 +65,7 @@ bool Joystick::isButtonPressed(unsigned int joystick, unsigned int button)
 ////////////////////////////////////////////////////////////
 float Joystick::getAxisPosition(unsigned int joystick, Axis axis)
 {
-    return priv::JoystickManager::getInstance().getState(joystick).axes[static_cast<int>(axis)];
+    return priv::JoystickManager::getInstance().getState(joystick).axes[axis];
 }
